Restructure alerts
