+++ src/core/Alerts/Alert.cs
@@ -87,4 +87,18 @@ namespace core.Alerts
             }
         }
     }
+
+    public struct AlertPricePoint
+    {
+        public AlertPricePoint(Guid id, string description, double value)
+        {
+            Id = id;
+            Description = description;
+            Value = value;
+        }
+
+        public Guid Id { get; }
+        public string Description { get; }
+        public double Value { get; }
+    }
 }
