--- src/core/Alerts/AlertPricePointAdded.cs
@@ -1,17 +1,0 @@
-using System;
-using core.Shared;
-using MediatR;
-
-namespace core.Alerts
-{
-    internal class AlertPricePointAdded : AggregateEvent
-    {
-        public AlertPricePointAdded(Guid id, Guid aggregateId, DateTimeOffset when, double value)
-            : base(id, aggregateId, when)
-        {
-            this.Value = value;
-        }
-
-        public double Value { get; }
-    }
-}
