+++ src/core/Alerts/Handlers/Clear.cs
