+++ src/core/Alerts/Handlers/Delete.cs
